Add files via upload
The Data Flow Diagram for the HouseholdHarmonizer is the conceptual design.  Right now I'm thinking that this will be a Swift project using the CoreML framework.  I believe that the speech detection and speech recognition models probably already exist.  The voice detection model that will be trained by the family instance of the application probably also exists.  The key piece of this technology that we'll need to create and train is the statement / question correlation model.

The application as envisioned may be too creepy for my original intent (people don't want Alexa listening in on their families).  However variants on this project will include:

- listening in on mediations to evaluate the use of open-ended questions and other tools
- listening in on lawyers to evaluate how often they are surprised by answers to their questions
- listening in on HR / interviews or evaluating listening skills in general